requirements.txt
Flask==2.3.3
dnspython==2.4.2
python-whois==0.8.0
requests==2.31.0
leakcheck==2.0.0
leakcheck